更新 Model 与 Animation 部分

+++ src/main/java/io/github/tt432/eyelib/util/math/MathE.java
@@ -24,7 +24,7 @@ public class MathE {
         return a + (b - a) * weight;
     }
 
-    public static double notZero(double value, double defaultValue) {
+    public static float notZero(float value, float defaultValue) {
         return value == 0 ? defaultValue : value;
     }
